LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/unloading-documents/unload/002-expected.txt

CONSOLE MESSAGE: TypeError: parent.do_test is not a function. (In 'parent.do_test()', 'parent.do_test' is undefined)

PASS pagehide event properties
